100Hz experiment
Fixed failure to update biases in RbmStack CPU version.
Added weight and bias magnitude reporting to RbmStack CPU version.
Added weight decay to RbmStack CPU version. This seems to be very
important for purposes of keeping the weights under control.
Added fixes above to GPU RbmStack version.
Tested GPU RbmStack with similar results.
Modified the SequenceDecimatingNetwork class so that it can perform
autoencoding in order to check the RMSE of the sequence decimating
network after pretraining. Based on these results using small networks
[(16,128),(2,128),(2,128)] I can see that the pretraining is somewhat
effective. Furthermore, the backpropagation training seems to be
somewhat effective in training a sequence decimating autoencoder with
RMSE less than the standard deviation of the input, at least for small
networks.
However, the backpropagation training doesn’t seem to improve the
network at all.
Training with [(16,128),(2,128),(2,128),(2,128),(2,1)] and zeroing
patient 1 and 2 because of nan, my submission AUC = 0.44683. Submitting
1-my predicted values did in fact negate this score with AUC = 0.55317,
yielding a position of 340 on the leaderboard.

Now running an experiemnt with 100Hz data and detrending disabled.